Restructured project and added multi-staging
