+++ .github/workflows/push.yml
@@ -21,5 +21,5 @@ jobs:
         pip install -r dependencies/requirements.txt
     - name: Analysing the code with pylint
       run: |
-        pylint v1 --rcfile=dependencies/app.pylintrc --fail-under=7
+        pylint src --rcfile=dependencies/app.pylintrc --fail-under=7
         pylint tests --rcfile=dependencies/tests.pylintrc --fail-under=8
