+++ .gitignore
@@ -107,6 +107,7 @@ celerybeat.pid
 *.sage.py
 
 # Environments
+.DS_Store
 .env
 .venv
 env/
