+++ dependencies/requirements.txt
@@ -66,3 +66,5 @@ watchfiles==0.15.0
 websockets==10.3
 wrapt==1.14.1
 yarl==1.7.2
+pre-commit==2.19.0
+pylint==2.14.4
