+++ docker-compose.yml
@@ -19,7 +19,7 @@ services:
   ports:
   - "5433:5432"
   env_file:
-  - ./configs/database-cfg.env
+  - src/configs/database-cfg.env
   volumes:
   - postgres:/var/lib/postgresql/data
