+++ src/configs/database-cfg-example.env
